Refactor into 3 files
